finished function edit experiment with goals and working on local host with postman
