+++ controller/experiment.controller.js
@@ -56,13 +56,6 @@ const getExperimentsByDate = async (req, res) => {
 }
 
 
-const updateExperimentsByID = async (req, res) => {
-    if (!req.params.experimentId) throw new PropertyNotFound("experimentId");
-    const experimentID = req.params.experimentId;
-    const result = await ExperimentRepository.update(experimentID, req.body)
-    if (!result) throw new ServerUnableError("updateExperimentsByID")
-    res.status(200).json(result);
-}
 
 const deleteExperimentsByID = async (req, res) => {
     if (!req.params.experimentId) throw new PropertyNotFound("experimentId");
@@ -93,7 +86,6 @@ module.exports = {
     getExperimentsAB,
     getExperimentsFF,
     getExperimentsByDate,
-    updateExperimentsByID,
     deleteExperimentsByID,
     addGoalToExperiment,
     removeGoalFromExperiment
