+++ router/experiment.router.js
@@ -1,9 +1,6 @@
 const express = require('express');
-
 const experimentController = require('../controller/experiment.controller');
-
-const newExperimentController = require ('../controller/experiment.creation.controller');
-
+const externalController = require('../controller/external.controller')
 const experimentRouter = new express.Router();
 
 experimentRouter.get('/', experimentController.getAllExperiments);
@@ -14,13 +11,13 @@ experimentRouter.get('/date', experimentController.getExperimentsByDate);
 experimentRouter.get('/:experimentId', experimentController.getExperimentById);
 
 
-experimentRouter.put('/:experimentId', experimentController.updateExperimentsByID);
+experimentRouter.put('/:experimentId', externalController.updateExperimentWithGoalsByExpID);
 experimentRouter.put('/:experimentId/:goalId', experimentController.addGoalToExperiment);
 
 experimentRouter.delete('/:experimentId/:goalId', experimentController.removeGoalFromExperiment);
 experimentRouter.delete('/:experimentId', experimentController.deleteExperimentsByID);
 
-experimentRouter.post('/new', newExperimentController.createExperimentWithGoals);
+experimentRouter.post('/new', externalController.createExperimentWithGoals);
 
 module.exports = {
     experimentRouter
